[android] fixed calculation of the animationDuration in the MyLocationView

+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/widgets/MyLocationView.java
@@ -999,7 +999,7 @@ public class MyLocationView extends View {
       }
 
       // updateLatLng timestamp
-      float previousUpdateTimeStamp = locationUpdateTimestamp;
+      long previousUpdateTimeStamp = locationUpdateTimestamp;
       locationUpdateTimestamp = SystemClock.elapsedRealtime();
 
       // calculate animation duration
@@ -1007,7 +1007,7 @@ public class MyLocationView extends View {
       if (previousUpdateTimeStamp == 0) {
         animationDuration = 0;
       } else {
-        animationDuration = (int) ((locationUpdateTimestamp - previousUpdateTimeStamp) * 1.1f)
+        animationDuration = (int) (((float)(locationUpdateTimestamp - previousUpdateTimeStamp)) * 1.1f)
         /*make animation slightly longer*/;
       }
